[key reuse] remove alternate simple implementation
